refacto, externalize configuration
