+++ views/home.ejs
@@ -34,6 +34,11 @@
                                         Connecté en tant que: <%= userEmail %>
                                     </a>
                                 </li>
+                                    <li>
+                                        <a class="fr-link fr-fi-add-circle-line" href="/logout">
+                                            Déconnexion
+                                        </a>
+                                    </li>
                                 <% } else {%>
                                 <li>
                                     non connecté
